Add websocket api in frontend
