+++ frontend/src/components/Waiter/Waiter.css
@@ -62,6 +62,11 @@
   animation-fill-mode: forwards;
   animation-duration: 200ms;
 }
+.box .buttonBox .reset:hover {
+  animation-name: fail;
+  animation-fill-mode: forwards;
+  animation-duration: 200ms;
+}
 @keyframes fail {
   100% {
     cursor: pointer;
